requirements.txt
flask
opencv-python-headless==4.8.0.76
ultralytics==8.0.142
Pillow==9.5.0
gunicorn
